Removed old logging system
