+++ src/cartodb.js
@@ -22,7 +22,7 @@ cdb.SQL = require('./api/sql');
 cdb.createVis = require('./api/create-vis');
 
 // log cartodb.js version
-var logger = require('./core/log');
+var logger = require('cdb.log');
 logger.log('cartodb.js ' + cdb.VERSION);
 
 module.exports = cdb;
