--- src/cdb.errors.js
@@ -1,2 +1,0 @@
-var ErrorList = require('./core/log/error-list');
-module.exports = new ErrorList();
